chore: Move update invite endpoint to endpoints section of app
