+++ tests/endpoints/tests_invites.py
@@ -1,3 +1,4 @@
+import json
 from unittest.mock import patch
 
 from tests.endpoints.endpoint_testing import TestEndpoints
@@ -30,3 +31,24 @@ class TestInvites(TestEndpoints):
 
         self.assertEqual(response.status_code, 200)
         self.assertEqual(data, mock_invites_data)
+
+    @patch(
+        "canonicalwebteam.store_api.dashboard.Dashboard.update_store_invites"
+    )
+    def test_update_invite_status(self, mock_update_store_invites):
+        """Test successful invite status update"""
+        mock_invites_data = [
+            {"id": "invite-1", "status": "accepted"},
+            {"id": "invite-2", "status": "declined"},
+        ]
+
+        mock_update_store_invites.return_value = None
+
+        response = self.client.post(
+            "/api/store/test-store-id/invite/update",
+            data={"invites": json.dumps(mock_invites_data)},
+        )
+        data = response.get_json()
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(data["msg"], "Changes saved")
