+++ webapp/admin/views.py
@@ -69,31 +69,6 @@ def post_invite_members(store_id):
     return jsonify(res)
 
 
-@admin.route("/api/store/<store_id>/invite/update", methods=["POST"])
-@login_required
-@exchange_required
-def update_invite_status(store_id):
-    invites = json.loads(flask.request.form.get("invites"))
-
-    res = {}
-
-    try:
-        dashboard.update_store_invites(flask.session, store_id, invites)
-        res["msg"] = "Changes saved"
-    except StoreApiResponseErrorList as api_response_error_list:
-        msgs = [
-            f"{error.get('message', 'An error occurred')}"
-            for error in api_response_error_list.errors
-        ]
-
-        msgs = list(dict.fromkeys(msgs))
-
-        for msg in msgs:
-            flask.flash(msg, "negative")
-
-    return jsonify(res)
-
-
 # -------------------- FEATURED SNAPS AUTOMATION ------------------
 @admin.route("/admin/featured", methods=["POST"])
 @login_required
